rename import.h and remove includes
