+++ openrct2.vcxproj
@@ -271,7 +271,7 @@
     <ClInclude Include="src\platform\crash.h" />
     <ClInclude Include="src\platform\platform.h" />
     <ClInclude Include="src\rct1.h" />
-    <ClInclude Include="src\rct1\import.h" />
+    <ClInclude Include="src\rct1\Tables.h" />
     <ClInclude Include="src\rct1\S4Importer.h" />
     <ClInclude Include="src\rct2.h" />
     <ClInclude Include="src\ride\cable_lift.h" />
