+++ src/rct1/tables.cpp
@@ -1,7 +1,7 @@
 #include "../common.h"
 #include "../core/List.hpp"
 #include "../core/Util.hpp"
-#include "import.h"
+#include "Tables.h"
 
 extern "C"
 {
